Muestra datos en el perfil
