+++ src/hooks/useAuth.tsx
@@ -1,7 +1,9 @@
 import { View, Text } from 'react-native'
-import React, { useState } from 'react'
+import React, { useState, useEffect } from 'react'
 import auth from '@react-native-firebase/auth'
 import database from '@react-native-firebase/database'
+import { GoogleSignin } from '@react-native-google-signin/google-signin';
+
 
 
 const useAuth = () => {
@@ -13,6 +15,12 @@ const useAuth = () => {
     //const [celular, setCelular] = useState('')
     const [messageError, setMessageError] = useState('')
 
+    useEffect(() => {
+        GoogleSignin.configure({
+            webClientId: '18752727933-rloes7gmi3s84q9e7f2lbajs20h917l7.apps.googleusercontent.com',
+        });
+    }, []);
+    
     const Login = () => {
         auth()
         .signInWithEmailAndPassword(email, password)
@@ -20,6 +28,30 @@ const useAuth = () => {
             console.log(user);
         });
     };
+
+    const googleSignIn = async () => {
+        // Check if your device supports Google Play
+        await GoogleSignin.hasPlayServices({ showPlayServicesUpdateDialog: true });
+        // Get the users ID token
+        const { idToken } = await GoogleSignin.signIn();
+      
+        // Create a Google credential with the token
+        const googleCredential = auth.GoogleAuthProvider.credential(idToken);
+      
+        // Sign-in the user with the credential
+        return auth().signInWithCredential(googleCredential);
+      }
+
+      const handleSignOutGoogle = async () => {
+        try {
+          await GoogleSignin.revokeAccess();
+          await auth().signOut()
+          console.log('Sesion Cerrada');
+          
+        } catch (error) {
+          console.error(error);
+        }
+      };
 
     const onRegister = () => {
         console.log("Register");
@@ -74,6 +106,8 @@ const useAuth = () => {
     return {
         Login,
         onRegister,
+        googleSignIn,
+        handleSignOutGoogle,
         email, 
         password,
         confirmPass,
