+++ src/screens/PerfilScreen.tsx
@@ -6,13 +6,17 @@ import { TouchableOpacity } from 'react-native'
 import Icon from 'react-native-vector-icons/Ionicons';
 import { ScrollView } from 'react-native-gesture-handler'
 import auth from '@react-native-firebase/auth'
+import useAuth from '../hooks/useAuth'
 
 
 
 interface Props extends StackScreenProps<any, any>{};
 
+const user = auth().currentUser;
+
 export const PerfilScreen = ( {navigation}: Props ) => {
   const [ useData, setUserData] = useState({});
+  const { handleSignOutGoogle } = useAuth();
   return (
     <View>
       <View style={ styles.barra}>
@@ -21,24 +25,24 @@ export const PerfilScreen = ( {navigation}: Props ) => {
 
       <ScrollView style={ styles.globalMargin}>
         <View style={styles.container}>
-          <Image source={require('../assets/images/imgUser.png')} style={styles.image} />
-        
+          <Image source={{ uri: user?.photoURL }} style={styles.image} />
+            
         </View>
 
         <View>
-          <Text style={ styles.h3 }> Nombre de Usuario </Text>
+          <Text style={ styles.h3 }> { user?.displayName } </Text>
         </View>
 
         <View style={ styles.datosPerfil}>
           <Text style={ styles.h4 }> Correo electronico: 
             <Text 
-              style={ styles.h5 }> { useData.email } 
+              style={ styles.h5 }> { user?.email }
             </Text> {' '}
            </Text>
         </View>
 
         <View style={ styles.datosPerfil}>
-          <Text style={ styles.h4 }> Celular: <Text style={ styles.h5 }> 314 123 4567 </Text> </Text>
+          <Text style={ styles.h4 }> Celular: <Text style={ styles.h5 }> { user?.phoneNumber } 3141234567</Text> </Text>
         </View>
 
         <View style={ styles.datosPerfil}>
@@ -70,7 +74,7 @@ export const PerfilScreen = ( {navigation}: Props ) => {
         <View>
           <View style={ styles.ButtonCerrar}>
             <TouchableOpacity style={ styles.btnCerrar}
-              onPress={ () => auth().signOut()}
+              onPress={ handleSignOutGoogle}
             >
                 <Text style={ styles.btnTextCerrar }> Cerrar Sesión <Text style={ styles.texto }> <Icon name="exit-outline"
                 size={25}
